perf(finance): flip EBITDA forecast to read from fact_opex_distribution_mat
DistributionAwareOpexProvider.getDistributionAwareOpex no longer computes
distribution at request time. It now reads from two materialized tables:
  - settled months  → fact_opex_mat (unchanged)
  - unsettled months → fact_opex_distribution_mat (new, populated by the
    nightly OpexDistributionRefreshBatchlet)

The distribution algorithm moves wholesale to OpexDistributionRefreshService
as private helpers (computeDistributionForMonths is package-private so the
in-package parity test can still verify it). @CacheResult annotations on
the resource method and on computeDistributionForMonth are removed — the
DB is the cache now.

OpexDistributionRefreshParityIT was retargeted to call the refresh
service's new computeDistributionForMonths entry point rather than the
provider's removed singular variant.

Also reverts OpexDistributionRefreshBatchlet.onStart back to its original
'only when empty' check, now that the tx-timeout fix has been verified
on staging.

Eliminates the >30s cold-path that was causing the BFF timeout on
/executive-dashboard. p95 should drop from >30s to <500ms.
